+++ sdk/netapp/azure-mgmt-netapp/azure/mgmt/netapp/models/_models_py3.py
@@ -99,7 +99,7 @@ class ActiveDirectory(_serialization.Model):  # pylint: disable=too-many-instanc
     :ivar administrators: Users to be added to the Built-in Administrators active directory group.
      A list of unique usernames without domain specifier.
     :vartype administrators: list[str]
-    :ivar kdc_ip: kdc server IP addresses for the active directory machine. This optional parameter
+    :ivar kdc_ip: kdc server IP address for the active directory machine. This optional parameter
      is used only while creating kerberos volume.
     :vartype kdc_ip: str
     :ivar ad_name: Name of the active directory machine. This optional parameter is used only while
@@ -229,7 +229,7 @@ class ActiveDirectory(_serialization.Model):  # pylint: disable=too-many-instanc
         :keyword administrators: Users to be added to the Built-in Administrators active directory
          group. A list of unique usernames without domain specifier.
         :paramtype administrators: list[str]
-        :keyword kdc_ip: kdc server IP addresses for the active directory machine. This optional
+        :keyword kdc_ip: kdc server IP address for the active directory machine. This optional
          parameter is used only while creating kerberos volume.
         :paramtype kdc_ip: str
         :keyword ad_name: Name of the active directory machine. This optional parameter is used only
@@ -603,7 +603,7 @@ class BackupPolicy(TrackedResource):  # pylint: disable=too-many-instance-attrib
     :vartype location: str
     :ivar etag: A unique read-only string that changes whenever the resource is updated.
     :vartype etag: str
-    :ivar backup_policy_id: Backup Policy Resource ID.
+    :ivar backup_policy_id: Backup Policy GUID ID.
     :vartype backup_policy_id: str
     :ivar provisioning_state: Azure lifecycle management.
     :vartype provisioning_state: str
@@ -704,7 +704,7 @@ class BackupPolicyPatch(_serialization.Model):  # pylint: disable=too-many-insta
     :vartype type: str
     :ivar tags: Resource tags.
     :vartype tags: dict[str, str]
-    :ivar backup_policy_id: Backup Policy Resource ID.
+    :ivar backup_policy_id: Backup Policy GUID ID.
     :vartype backup_policy_id: str
     :ivar provisioning_state: Azure lifecycle management.
     :vartype provisioning_state: str
@@ -1440,6 +1440,30 @@ class CloudErrorBody(_serialization.Model):
         self.message = message
 
 
+class ClusterPeerCommandResponse(_serialization.Model):
+    """Information about cluster peering process.
+
+    :ivar peer_accept_command: A command that needs to be run on the external ONTAP to accept
+     cluster peering.  Will only be present if :code:`<code>clusterPeeringStatus</code>` is
+     :code:`<code>pending</code>`.
+    :vartype peer_accept_command: str
+    """
+
+    _attribute_map = {
+        "peer_accept_command": {"key": "peerAcceptCommand", "type": "str"},
+    }
+
+    def __init__(self, *, peer_accept_command: Optional[str] = None, **kwargs: Any) -> None:
+        """
+        :keyword peer_accept_command: A command that needs to be run on the external ONTAP to accept
+         cluster peering.  Will only be present if :code:`<code>clusterPeeringStatus</code>` is
+         :code:`<code>pending</code>`.
+        :paramtype peer_accept_command: str
+        """
+        super().__init__(**kwargs)
+        self.peer_accept_command = peer_accept_command
+
+
 class DailySchedule(_serialization.Model):
     """Daily Schedule properties.
 
@@ -1791,6 +1815,10 @@ class FilePathAvailabilityRequest(_serialization.Model):
     :ivar subnet_id: The Azure Resource URI for a delegated subnet. Must have the delegation
      Microsoft.NetApp/volumes. Required.
     :vartype subnet_id: str
+    :ivar availability_zone: The Azure Resource logical availability zone which is used within zone
+     mapping lookup for the subscription and region. The lookup will retrieve the physical zone
+     where volume is placed.
+    :vartype availability_zone: str
     """
 
     _validation = {
@@ -1801,19 +1829,25 @@ class FilePathAvailabilityRequest(_serialization.Model):
     _attribute_map = {
         "name": {"key": "name", "type": "str"},
         "subnet_id": {"key": "subnetId", "type": "str"},
+        "availability_zone": {"key": "availabilityZone", "type": "str"},
     }
 
-    def __init__(self, *, name: str, subnet_id: str, **kwargs: Any) -> None:
+    def __init__(self, *, name: str, subnet_id: str, availability_zone: Optional[str] = None, **kwargs: Any) -> None:
         """
         :keyword name: File path to verify. Required.
         :paramtype name: str
         :keyword subnet_id: The Azure Resource URI for a delegated subnet. Must have the delegation
          Microsoft.NetApp/volumes. Required.
         :paramtype subnet_id: str
+        :keyword availability_zone: The Azure Resource logical availability zone which is used within
+         zone mapping lookup for the subscription and region. The lookup will retrieve the physical zone
+         where volume is placed.
+        :paramtype availability_zone: str
         """
         super().__init__(**kwargs)
         self.name = name
         self.subnet_id = subnet_id
+        self.availability_zone = availability_zone
 
 
 class GetGroupIdListForLDAPUserRequest(_serialization.Model):
@@ -2876,6 +2910,34 @@ class OperationListResult(_serialization.Model):
         self.value = value
 
 
+class PeerClusterForVolumeMigrationRequest(_serialization.Model):
+    """Source Cluster properties for a cluster peer request.
+
+    All required parameters must be populated in order to send to server.
+
+    :ivar peer_ip_addresses: A list of IC-LIF IPs that can be used to connect to the On-prem
+     cluster. Required.
+    :vartype peer_ip_addresses: list[str]
+    """
+
+    _validation = {
+        "peer_ip_addresses": {"required": True, "min_items": 1},
+    }
+
+    _attribute_map = {
+        "peer_ip_addresses": {"key": "peerIpAddresses", "type": "[str]"},
+    }
+
+    def __init__(self, *, peer_ip_addresses: List[str], **kwargs: Any) -> None:
+        """
+        :keyword peer_ip_addresses: A list of IC-LIF IPs that can be used to connect to the On-prem
+         cluster. Required.
+        :paramtype peer_ip_addresses: list[str]
+        """
+        super().__init__(**kwargs)
+        self.peer_ip_addresses = peer_ip_addresses
+
+
 class PlacementKeyValuePairs(_serialization.Model):
     """Application specific parameters for the placement of volumes in the volume group.
 
@@ -3247,6 +3309,46 @@ class RelocateVolumeRequest(_serialization.Model):
         self.creation_token = creation_token
 
 
+class RemotePath(_serialization.Model):
+    """The full path to a volume that is to be migrated into ANF. Required for Migration volumes.
+
+    All required parameters must be populated in order to send to server.
+
+    :ivar external_host_name: The Path to a ONTAP Host. Required.
+    :vartype external_host_name: str
+    :ivar server_name: The name of a server on the ONTAP Host. Required.
+    :vartype server_name: str
+    :ivar volume_name: The name of a volume on the server. Required.
+    :vartype volume_name: str
+    """
+
+    _validation = {
+        "external_host_name": {"required": True},
+        "server_name": {"required": True},
+        "volume_name": {"required": True},
+    }
+
+    _attribute_map = {
+        "external_host_name": {"key": "externalHostName", "type": "str"},
+        "server_name": {"key": "serverName", "type": "str"},
+        "volume_name": {"key": "volumeName", "type": "str"},
+    }
+
+    def __init__(self, *, external_host_name: str, server_name: str, volume_name: str, **kwargs: Any) -> None:
+        """
+        :keyword external_host_name: The Path to a ONTAP Host. Required.
+        :paramtype external_host_name: str
+        :keyword server_name: The name of a server on the ONTAP Host. Required.
+        :paramtype server_name: str
+        :keyword volume_name: The name of a volume on the server. Required.
+        :paramtype volume_name: str
+        """
+        super().__init__(**kwargs)
+        self.external_host_name = external_host_name
+        self.server_name = server_name
+        self.volume_name = volume_name
+
+
 class Replication(_serialization.Model):
     """Replication properties.
 
@@ -3330,6 +3432,9 @@ class ReplicationObject(_serialization.Model):
     :vartype replication_schedule: str or ~azure.mgmt.netapp.models.ReplicationSchedule
     :ivar remote_volume_resource_id: The resource ID of the remote volume. Required.
     :vartype remote_volume_resource_id: str
+    :ivar remote_path: The full path to a volume that is to be migrated into ANF. Required for
+     Migration volumes.
+    :vartype remote_path: ~azure.mgmt.netapp.models.RemotePath
     :ivar remote_volume_region: The remote region for the other end of the Volume Replication.
     :vartype remote_volume_region: str
     """
@@ -3344,6 +3449,7 @@ class ReplicationObject(_serialization.Model):
         "endpoint_type": {"key": "endpointType", "type": "str"},
         "replication_schedule": {"key": "replicationSchedule", "type": "str"},
         "remote_volume_resource_id": {"key": "remoteVolumeResourceId", "type": "str"},
+        "remote_path": {"key": "remotePath", "type": "RemotePath"},
         "remote_volume_region": {"key": "remoteVolumeRegion", "type": "str"},
     }
 
@@ -3353,6 +3459,7 @@ class ReplicationObject(_serialization.Model):
         remote_volume_resource_id: str,
         endpoint_type: Optional[Union[str, "_models.EndpointType"]] = None,
         replication_schedule: Optional[Union[str, "_models.ReplicationSchedule"]] = None,
+        remote_path: Optional["_models.RemotePath"] = None,
         remote_volume_region: Optional[str] = None,
         **kwargs: Any
     ) -> None:
@@ -3365,6 +3472,9 @@ class ReplicationObject(_serialization.Model):
         :paramtype replication_schedule: str or ~azure.mgmt.netapp.models.ReplicationSchedule
         :keyword remote_volume_resource_id: The resource ID of the remote volume. Required.
         :paramtype remote_volume_resource_id: str
+        :keyword remote_path: The full path to a volume that is to be migrated into ANF. Required for
+         Migration volumes.
+        :paramtype remote_path: ~azure.mgmt.netapp.models.RemotePath
         :keyword remote_volume_region: The remote region for the other end of the Volume Replication.
         :paramtype remote_volume_region: str
         """
@@ -3373,6 +3483,7 @@ class ReplicationObject(_serialization.Model):
         self.endpoint_type = endpoint_type
         self.replication_schedule = replication_schedule
         self.remote_volume_resource_id = remote_volume_resource_id
+        self.remote_path = remote_path
         self.remote_volume_region = remote_volume_region
 
 
@@ -4340,6 +4451,30 @@ class SubvolumesList(_serialization.Model):
         self.next_link = next_link
 
 
+class SvmPeerCommandResponse(_serialization.Model):
+    """Information about svm peering process.
+
+    :ivar svm_peering_command: A command that needs to be run on the external ONTAP to accept svm
+     peering.  Will only be present if :code:`<code>svmPeeringStatus</code>` is
+     :code:`<code>pending</code>`.
+    :vartype svm_peering_command: str
+    """
+
+    _attribute_map = {
+        "svm_peering_command": {"key": "svmPeeringCommand", "type": "str"},
+    }
+
+    def __init__(self, *, svm_peering_command: Optional[str] = None, **kwargs: Any) -> None:
+        """
+        :keyword svm_peering_command: A command that needs to be run on the external ONTAP to accept
+         svm peering.  Will only be present if :code:`<code>svmPeeringStatus</code>` is
+         :code:`<code>pending</code>`.
+        :paramtype svm_peering_command: str
+        """
+        super().__init__(**kwargs)
+        self.svm_peering_command = svm_peering_command
+
+
 class SystemData(_serialization.Model):
     """Metadata pertaining to creation and last modification of the resource.
 
@@ -4561,9 +4696,14 @@ class Volume(TrackedResource):  # pylint: disable=too-many-instance-attributes
     :ivar subnet_id: The Azure Resource URI for a delegated subnet. Must have the delegation
      Microsoft.NetApp/volumes. Required.
     :vartype subnet_id: str
-    :ivar network_features: Network features available to the volume, or current state of update.
-     Known values are: "Basic", "Standard", "Basic_Standard", and "Standard_Basic".
+    :ivar network_features: The original value of the network features type available to the volume
+     at the time it was created. Known values are: "Basic", "Standard", "Basic_Standard", and
+     "Standard_Basic".
     :vartype network_features: str or ~azure.mgmt.netapp.models.NetworkFeatures
+    :ivar effective_network_features: The effective value of the network features type available to
+     the volume, or current effective state of update. Known values are: "Basic", "Standard",
+     "Basic_Standard", and "Standard_Basic".
+    :vartype effective_network_features: str or ~azure.mgmt.netapp.models.NetworkFeatures
     :ivar network_sibling_set_id: Network Sibling Set ID for the the group of volumes sharing
      networking resources.
     :vartype network_sibling_set_id: str
@@ -4721,6 +4861,7 @@ class Volume(TrackedResource):  # pylint: disable=too-many-instance-attributes
         "provisioning_state": {"readonly": True},
         "baremetal_tenant_id": {"readonly": True},
         "subnet_id": {"required": True},
+        "effective_network_features": {"readonly": True},
         "network_sibling_set_id": {
             "readonly": True,
             "max_length": 36,
@@ -4765,6 +4906,7 @@ class Volume(TrackedResource):  # pylint: disable=too-many-instance-attributes
         "baremetal_tenant_id": {"key": "properties.baremetalTenantId", "type": "str"},
         "subnet_id": {"key": "properties.subnetId", "type": "str"},
         "network_features": {"key": "properties.networkFeatures", "type": "str"},
+        "effective_network_features": {"key": "properties.effectiveNetworkFeatures", "type": "str"},
         "network_sibling_set_id": {"key": "properties.networkSiblingSetId", "type": "str"},
         "storage_to_network_proximity": {"key": "properties.storageToNetworkProximity", "type": "str"},
         "mount_targets": {"key": "properties.mountTargets", "type": "[MountTargetProperties]"},
@@ -4889,8 +5031,9 @@ class Volume(TrackedResource):  # pylint: disable=too-many-instance-attributes
         :keyword subnet_id: The Azure Resource URI for a delegated subnet. Must have the delegation
          Microsoft.NetApp/volumes. Required.
         :paramtype subnet_id: str
-        :keyword network_features: Network features available to the volume, or current state of
-         update. Known values are: "Basic", "Standard", "Basic_Standard", and "Standard_Basic".
+        :keyword network_features: The original value of the network features type available to the
+         volume at the time it was created. Known values are: "Basic", "Standard", "Basic_Standard", and
+         "Standard_Basic".
         :paramtype network_features: str or ~azure.mgmt.netapp.models.NetworkFeatures
         :keyword volume_type: What type of volume is this. For destination volumes in Cross Region
          Replication, set type to DataProtection.
@@ -5003,6 +5146,7 @@ class Volume(TrackedResource):  # pylint: disable=too-many-instance-attributes
         self.baremetal_tenant_id = None
         self.subnet_id = subnet_id
         self.network_features = network_features
+        self.effective_network_features = None
         self.network_sibling_set_id = None
         self.storage_to_network_proximity = None
         self.mount_targets = None
@@ -5376,9 +5520,14 @@ class VolumeGroupVolumeProperties(_serialization.Model):  # pylint: disable=too-
     :ivar subnet_id: The Azure Resource URI for a delegated subnet. Must have the delegation
      Microsoft.NetApp/volumes. Required.
     :vartype subnet_id: str
-    :ivar network_features: Network features available to the volume, or current state of update.
-     Known values are: "Basic", "Standard", "Basic_Standard", and "Standard_Basic".
+    :ivar network_features: The original value of the network features type available to the volume
+     at the time it was created. Known values are: "Basic", "Standard", "Basic_Standard", and
+     "Standard_Basic".
     :vartype network_features: str or ~azure.mgmt.netapp.models.NetworkFeatures
+    :ivar effective_network_features: The effective value of the network features type available to
+     the volume, or current effective state of update. Known values are: "Basic", "Standard",
+     "Basic_Standard", and "Standard_Basic".
+    :vartype effective_network_features: str or ~azure.mgmt.netapp.models.NetworkFeatures
     :ivar network_sibling_set_id: Network Sibling Set ID for the the group of volumes sharing
      networking resources.
     :vartype network_sibling_set_id: str
@@ -5532,6 +5681,7 @@ class VolumeGroupVolumeProperties(_serialization.Model):  # pylint: disable=too-
         "provisioning_state": {"readonly": True},
         "baremetal_tenant_id": {"readonly": True},
         "subnet_id": {"required": True},
+        "effective_network_features": {"readonly": True},
         "network_sibling_set_id": {
             "readonly": True,
             "max_length": 36,
@@ -5573,6 +5723,7 @@ class VolumeGroupVolumeProperties(_serialization.Model):  # pylint: disable=too-
         "baremetal_tenant_id": {"key": "properties.baremetalTenantId", "type": "str"},
         "subnet_id": {"key": "properties.subnetId", "type": "str"},
         "network_features": {"key": "properties.networkFeatures", "type": "str"},
+        "effective_network_features": {"key": "properties.effectiveNetworkFeatures", "type": "str"},
         "network_sibling_set_id": {"key": "properties.networkSiblingSetId", "type": "str"},
         "storage_to_network_proximity": {"key": "properties.storageToNetworkProximity", "type": "str"},
         "mount_targets": {"key": "properties.mountTargets", "type": "[MountTargetProperties]"},
@@ -5697,8 +5848,9 @@ class VolumeGroupVolumeProperties(_serialization.Model):  # pylint: disable=too-
         :keyword subnet_id: The Azure Resource URI for a delegated subnet. Must have the delegation
          Microsoft.NetApp/volumes. Required.
         :paramtype subnet_id: str
-        :keyword network_features: Network features available to the volume, or current state of
-         update. Known values are: "Basic", "Standard", "Basic_Standard", and "Standard_Basic".
+        :keyword network_features: The original value of the network features type available to the
+         volume at the time it was created. Known values are: "Basic", "Standard", "Basic_Standard", and
+         "Standard_Basic".
         :paramtype network_features: str or ~azure.mgmt.netapp.models.NetworkFeatures
         :keyword volume_type: What type of volume is this. For destination volumes in Cross Region
          Replication, set type to DataProtection.
@@ -5814,6 +5966,7 @@ class VolumeGroupVolumeProperties(_serialization.Model):  # pylint: disable=too-
         self.baremetal_tenant_id = None
         self.subnet_id = subnet_id
         self.network_features = network_features
+        self.effective_network_features = None
         self.network_sibling_set_id = None
         self.storage_to_network_proximity = None
         self.mount_targets = None