+++ sdk/netapp/azure-mgmt-netapp/generated_samples/volumes_finalize_external_replication.py
@@ -15,7 +15,7 @@ from azure.mgmt.netapp import NetAppManagementClient
     pip install azure-identity
     pip install azure-mgmt-netapp
 # USAGE
-    python accounts_create_or_update.py
+    python volumes_finalize_external_replication.py
 
     Before run the sample, please set the values of the client ID, tenant ID and client secret
     of the AAD application as environment variables: AZURE_CLIENT_ID, AZURE_TENANT_ID,
@@ -30,14 +30,14 @@ def main():
         subscription_id="D633CC2E-722B-4AE1-B636-BBD9E4C60ED9",
     )
 
-    response = client.accounts.begin_create_or_update(
+    client.volumes.begin_finalize_external_replication(
         resource_group_name="myRG",
         account_name="account1",
-        body={"location": "eastus", "properties": {}},
+        pool_name="pool1",
+        volume_name="volume1",
     ).result()
-    print(response)
 
 
-# x-ms-original-file: specification/netapp/resource-manager/Microsoft.NetApp/stable/2024-07-01/examples/Accounts_CreateOrUpdate.json
+# x-ms-original-file: specification/netapp/resource-manager/Microsoft.NetApp/stable/2024-07-01/examples/Volumes_FinalizeExternalReplication.json
 if __name__ == "__main__":
     main()